Revert "Merge remote-tracking branch 'origin/master'"
This reverts commit 23c924626c2fa07cb170dd89e932923265384050, reversing
changes made to c2a35e335638466019eaa098d753bdb7da5528b2.
