+++ Cozyfications/bot/core/bot.py
@@ -7,8 +7,7 @@ import discord
 from aiohttp import ClientSession
 
 from Cozyfications import secrets
-from callbacks import Callbacks
-from views import ConfirmDialog
+from Cozyfications.bot import core
 
 
 class Cozyfications(discord.Bot):
@@ -35,9 +34,9 @@ class Cozyfications(discord.Bot):
         self.new_subscriptions: int = 0
         self.delete_subscriptions: int = 0
         self.subscriptions: dict[str, callable] = {
-            "channel.update": Callbacks.update,
-            "stream.online": Callbacks.online,
-            "stream.offline": Callbacks.offline
+            "channel.update": core.Callbacks.update,
+            "stream.online": core.Callbacks.online,
+            "stream.offline": core.Callbacks.offline
         }
 
         self.port: int = 6001
@@ -72,7 +71,7 @@ class Cozyfications(discord.Bot):
             bot_token=self.http.token,
         )
 
-        self.add_view(view=ConfirmDialog())
+        self.add_view(view=core.ConfirmDialog())
 
         msg: str = f"""{self.user.name} is online now!
             BotID: {self.user.id}
