+++ Cozyfications/bot/core/embeds.py
@@ -2,8 +2,8 @@ import datetime
 
 import discord
 
-from __init__ import Cog
-from bot import Cozyfications
+from .__init__ import Cog
+from .bot import Cozyfications
 
 
 class Embed(discord.Embed):
@@ -255,6 +255,47 @@ class HelpSelectEmbed(CozyficationsEmbed):
         self.description: str = "\n".join(
             f"`/{command.qualified_name}`: {command.description}"
             for command in self.cog.walk_commands()
+        )
+
+
+class HelpSelect(discord.ui.Select):
+    """Represents a custom PyCord UI help select menu."""
+
+    def __init__(self, *, bot: Cozyfications, cog: Cog) -> None:
+        """Initialises a new help select menu.
+
+        Parameters
+        ----------
+        bot: :class:`Cozyfications`
+            The bot instance.
+        cog: :class:`Cog`
+            The cog instance."""
+        self.bot: Cozyfications = bot
+        self.cog: Cog = cog
+        super().__init__(
+            placeholder="Choose a category",
+            options=[
+                discord.SelectOption(
+                    label=cog_name,
+                    description=cog.__doc__,
+                )
+                for cog_name, cog in self.cog.bot.cogs.items()
+                if cog.__cog_commands__ and cog_name not in ["Help"]
+            ],
+        )
+
+    async def callback(self, interaction: discord.Interaction) -> None:
+        """Handles the callback for the help select menu.
+
+        Parameters
+        ----------
+        interaction: :class:`discord.Interaction`
+            The interaction instance."""
+        cog = self.cog.bot.get_cog(self.values[0])
+        embed = HelpSelectEmbed(bot=self.bot, cog=cog)
+        await interaction.response.send_message(
+            embed=embed,
+            ephemeral=True,
         )
